cleanup old proxy server
