+++ mitmproxy/net/http/http1/__init__.py
@@ -1,7 +1,6 @@
 from .read import (
-    read_request, read_request_head,
-    read_response, read_response_head,
-    read_body,
+    read_request_head,
+    read_response_head,
     connection_close,
     expected_http_body_size,
 )
@@ -13,9 +12,8 @@ from .assemble import (
 
 
 __all__ = [
-    "read_request", "read_request_head",
-    "read_response", "read_response_head",
-    "read_body",
+    "read_request_head",
+    "read_response_head",
     "connection_close",
     "expected_http_body_size",
     "assemble_request", "assemble_request_head",
