+++ mitmproxy/tools/dump.py
@@ -1,7 +1,7 @@
 from mitmproxy import addons
 from mitmproxy import options
 from mitmproxy import master
-from mitmproxy.addons import dumper, termlog, termstatus, keepserving, readfile
+from mitmproxy.addons import dumper, termlog, keepserving, readfile
 
 
 class ErrorCheck:
@@ -24,7 +24,7 @@ class DumpMaster(master.Master):
         super().__init__(options)
         self.errorcheck = ErrorCheck()
         if with_termlog:
-            self.addons.add(termlog.TermLog(), termstatus.TermStatus())
+            self.addons.add(termlog.TermLog())
         self.addons.add(*addons.default_addons())
         if with_dumper:
             self.addons.add(dumper.Dumper())
