+++ test/mitmproxy/addons/test_disable_h2c.py
@@ -1,9 +1,6 @@
-import io
-
 from mitmproxy.addons import disable_h2c
 from mitmproxy.exceptions import Kill
-from mitmproxy.net.http import http1
-from mitmproxy.test import taddons
+from mitmproxy.test import taddons, tutils
 from mitmproxy.test import tflow
 
 
@@ -28,9 +25,12 @@ class TestDisableH2CleartextUpgrade:
             a = disable_h2c.DisableH2C()
             tctx.configure(a)
 
-            b = io.BytesIO(b"PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n")
             f = tflow.tflow()
-            f.request = http1.read_request(b)
+            f.request = tutils.treq(
+                method=b"PRI",
+                path=b"*",
+                http_version=b"HTTP/2.0",
+            )
             f.intercept()
 
             a.request(f)
