--- test/mitmproxy/addons/test_termstatus.py
@@ -1,18 +1,0 @@
-import pytest
-
-from mitmproxy import proxy
-from mitmproxy.addons import termstatus
-from mitmproxy.test import taddons
-
-
-@pytest.mark.asyncio
-async def test_configure():
-    ts = termstatus.TermStatus()
-    with taddons.context() as ctx:
-        ctx.master.server = proxy.DummyServer()
-        ctx.master.server.bound = True
-        ctx.configure(ts, server=False)
-        ts.running()
-        ctx.configure(ts, server=True)
-        ts.running()
-        await ctx.master.await_log("server listening")
