+++ test/mitmproxy/tools/test_cmdline.py
@@ -10,7 +10,7 @@ def test_common():
     opts = options.Options()
     cmdline.common_options(parser, opts)
     args = parser.parse_args(args=[])
-    assert main.process_options(parser, opts, args)
+    main.process_options(parser, opts, args)
 
 
 def test_mitmproxy():
